refactor(runtime-tags): collapse the duplicated for-type and statement tag definitions
