+++ packages/runtime-tags/src/translator/core/static.ts
@@ -1,31 +1,3 @@
-import { types as t } from "@marko/compiler";
-import { parseStatements, type Tag } from "@marko/compiler/babel-utils";
-export default {
-  parse(tag) {
-    const {
-      node,
-      hub: { file },
-    } = tag;
-    const rawValue = node.rawValue!;
-    const code = rawValue.replace(/^static\s*/, "");
-    const start = node.start! + (rawValue.length - code.length);
-    let body = parseStatements(file, code, start, start + code.length);
-    if (body.length === 1 && t.isBlockStatement(body[0])) {
-      body = body[0].body;
-    }
+import { createStatementTag } from "../util/statement-tag";
 
-    tag.replaceWith(t.markoScriptlet(body, true));
-  },
-  parseOptions: {
-    statement: true,
-    rawOpenTag: true,
-  },
-  autocomplete: [
-    {
-      displayText: "static <statement>",
-      description:
-        "A JavaScript statement which is only evaluated once your template is loaded.",
-      descriptionMoreURL: "https://markojs.com/docs/syntax/#static-javascript",
-    },
-  ],
-} as Tag;
+export default createStatementTag("static");
